Move getScriptArgs into utils and document the text domain options
